extract common methods to abstract parser class
