+++ checkov/sast/checks_infra/base_registry.py
@@ -11,8 +11,8 @@ from checkov.common.bridgecrew.check_type import CheckType
 from checkov.sast.checks_infra.base_check import BaseSastCheck
 from checkov.common.checks.base_check_registry import BaseCheckRegistry
 from checkov.runner_filter import RunnerFilter
-from checkov.sast.checks_infra.checks_parser_v01 import SastCheckParserV01
-from checkov.sast.checks_infra.checks_parser_v02 import SastCheckParserV02
+from checkov.sast.checks_infra.check_parser.parser_v01 import SastCheckParserV01
+from checkov.sast.checks_infra.check_parser.parser_v02 import SastCheckParserV02
 from checkov.sast.consts import SastLanguages, BqlVersion, get_bql_version_from_string
 from checkov.common.checks_infra.registry import CHECKS_POSSIBLE_ENDING
